Add an abstraction layer when handling hashes
This change allows the incoming Python 3 port to more easily treat
the hashes, which are sent by the client with their hexdigest represetantion
(which is a unicode/str), as str instead of bytes.

It also simplifies the bytes() call that is needed on every memoryview, while
still being backward compatible.

Unify hash encoding function into a helper in the utils module