Unify calculation of put content request params
There is a lot (A LOT) of code duplication when calculating the parameters to
call the `put content` request. This branch proposes a helper to do that
calculation in a simple manner and adapts tests to use it.

The tests for `content` (test_content.py) were pretty much re-written using
inlineCallbacks and this new helper, producing a much smaller and simpler file.